Compiles on windows
