+++ pkg/security/probe/opts_windows.go
@@ -3,9 +3,6 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build linux
-// +build linux
-
 package probe
 
 import (
@@ -14,8 +11,6 @@ import (
 
 // Opts defines some probe options
 type Opts struct {
-	// DontDiscardRuntime do not discard the runtime. Mostly used by functional tests
-	DontDiscardRuntime bool
 	// StatsdClient to be used for probe stats
 	StatsdClient statsd.ClientInterface
 }
